Renamed stackStatusChecker to statusChecker
