+++ code/workspaces/infrastructure-api/src/kubeWatcher/statusChecker.js
